Feature transaction history ruth

+++ app/src/main/java/com/decagonhq/clads/data/domain/client/ClientReg.kt
@@ -3,7 +3,7 @@ package com.decagonhq.clads.data.domain.client
 // for first step registration
 data class ClientReg(
     val id: Int? = null,
-    val artisanId: Int? = null,
+    val artisanId: Long? = null,
     val email: String,
     val fullName: String,
     val gender: String,
